Add files via upload
I suppose, I should have been commiting directly to my branch from the start and then placing a pull request to merge, rather than continuing to make new branches. 

Notes: 
XGB with range of depths (1, 10,1) wall time of 41.9, but accuracy on testing of 0.854

XGB with [5,7,9] wall time of 19.3 s, but accuarcy on testing of 0.853

SVC model on RBF kernel C=0.1, gamma=0.1 has a big wall time of 6min 26s and yields accuracy of 0.852

XGB seems like best model on this data set, highly accurate and low wall time. MSE = 0.188 and accuracy = 0.854

Feature importances for XGB should be put into slides, maybe compare to some of the slightly less good models?

Very intersting to me is how Random Forests importance of features is very different from XGB.

Not sure how much we should read into bagging. It's best train MSE was 0.076, but seemingly it got "lucky" and has a test MSE of 0.195? If we had more time, it would be interesting to investigate why this is or swap out test/train splits.

Another thing I find interesting, Decision tree doesn't like to increase the depth of the tree. It keeps finding the minimum depth to be optimal.

Finally, but really firstly, linear regression of all types seems pitiful on this data set. The only VIF on a feature that isn't below 10 is Road surface type, but no other model thinks this feature is super important (ranking it near the bottom of feature importance, but not usally the bottom).  

What are your thoughts on calling the code here, and starting on the presentation? I think code is about as fine tuned as we have time for. 85.4% accuracy seems decent-ish.