Testing XGB and SVC_rbf with og (uncleaned data)
I got curious and wanted to check what the results would be if we used XGB model on just the raw data with nothing done to it. It winds up being pretty good even with all of the null data, colinear features, and mostly useless features: XGB Training Accuracy with og data is: 0.9528; XGB Testing Accuracy with og data is: 0.8677; Wall time: 1min 52s
 as compared to our other results: Training Accuracy is: 0.8405; Testing Accuracy is: 0.854; Wall time: 40.2 s

I don't think we need to merge this into master branch. It can stay as this branch since we are done with the code. I just really wanted to know the answer to the professors question/statement about XGB probably being able to just handle the raw data up to like 100k rows. 